rpc: move test-only code to tests
